chore(instrumentation-graphql): remove operation name from execute span
