+++ plugins/node/opentelemetry-instrumentation-graphql/src/utils.ts
@@ -17,12 +17,7 @@
 import type * as graphqlTypes from 'graphql';
 import * as api from '@opentelemetry/api';
 import { GraphQLObjectType } from 'graphql/type/definition';
-import {
-  AllowedOperationTypes,
-  createResolveSpanName,
-  SpanNames,
-  TokenKind,
-} from './enum';
+import { AllowedOperationTypes, SpanNames, TokenKind } from './enum';
 import { AttributeNames } from './enums/AttributeNames';
 import { OTEL_GRAPHQL_DATA_SYMBOL, OTEL_PATCHED_SYMBOL } from './symbols';
 import {
@@ -438,3 +433,16 @@ async function safeExecuteInTheMiddleAsync<T>(
     return result as T;
   }
 }
+
+export const createResolveSpanName = (fieldName: string): string => {
+  return `${SpanNames.RESOLVE}.${fieldName}`;
+};
+
+export const createExecuteSpanName = (
+  operationType: string | undefined
+): string => {
+  if (operationType) {
+    return `${SpanNames.EXECUTE}.${operationType}`;
+  }
+  return `${SpanNames.EXECUTE}`;
+};
